[ir_reader] Move tests to separate module
